refactor(backend): extract card managed image settlement
