+++ apps/backend/src/cards/generatedImageAppend.test.ts
@@ -7,7 +7,7 @@ import {
   appendManagedImageToCardSideInExecutor,
   appendManagedImageToCardText,
   buildManagedImageMarkdownReference,
-} from "./index";
+} from "./managedImageSettlement";
 import type { AppendManagedImageToCardSideInput, CardMutationMetadata, CardRow, CardTextSide } from "./types";
 
 const testWorkspaceId = "22222222-2222-4222-8222-222222222222";
